Moved devops classes from installer package to hubturbo.updater package
